move AddCards into own file
